Introduce a full flow verification step
Introduce a new CI step that performs a full end-to-end flow
verification. It uses the binary compiled in this project to generate
the source code from the Protobuf files. It then uses the resulting code
in the example crate. It runs tests, creating a gRPC server and using a
gRPC client to talk to it and asserts that the entire flow works fully
as expected.