Handle Rust reserved keywords in the package
Protobuf allows to use values in the package name that are considered
Rust reserved keywords. As a result they need special handling. For
example to include the module named `type` it must be referenced as
`r#type`, but the actual module directory should be unchanged. Enable
handling of these keywords by escaping `pub mod ...` generated
statements.